decompose intersect node to pipeline

+++ be/src/exec/vectorized/intersect_hash_set.h
@@ -67,7 +67,7 @@ public:
     Status refine_intersect_row(RuntimeState* state, const ChunkPtr& chunkPtr, const std::vector<ExprContext*>& exprs,
                                 int hit_times);
 
-    void deserialize_to_columns(KeyVector& keys, const Columns& key_columns, int32_t batch_size);
+    void deserialize_to_columns(KeyVector& keys, const Columns& key_columns, size_t batch_size);
 
 private:
     void _serialize_columns(const ChunkPtr& chunkPtr, const std::vector<ExprContext*>& exprs, size_t chunk_size);
